様々な修正とバージョンの書き換え
