+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.1.3",
+  "version": "2.0.0",
   "dependencies": {
     "@discordjs/opus": "^0.8.0",
     "@discordjs/voice": "^0.11.0",
